MoE a16wmix: drop redundant gemm2 CSV resolver; move act.py to shared kernels dir
#2 (dead CSV code): after the fold, a16w4 gemm2 tiles are already carried by the
get_2stage_cfgs kernelName2 (parsed into tile_n/tile_k/b_nt/waves_per_eu/xcd_swizzle
by _flydsl_stage2_wrapper), so the stage2 impl re-reading the CSV via
resolve_a16w4_gemm2_config was redundant. Switch the a16w4 stage2 branch to the
passed tiles (same as a4w4/a8w4) and delete the a16w4 CSV family
(resolve_a16w4_gemm2_config, pick_a16w4_config, _load_a16w4_csv, _decode_a16w4_kname,
_default_a16w4_csv_path, kernelName regexes, FLYDSL_A16W4_TUNED_CSV env). csv_dispatch.py
is now heuristic-only -> renamed tile_heuristic.py.

#4: move act.py (SiLU/SiTUv2 helpers) out of moe_2stage_a16wmix/ to the shared
aiter/ops/flydsl/kernels/act.py so other FlyDSL MoE kernels can reuse it.

Validated (gfx950 kimik3): logits_diff a16w4 ~1.5e-5 / a8w4 ~2e-5; dedicated pytest
6/6; a16w4 e2e perf parity (tok4/8/256/16384 within noise; get_padded_M is identity
at those tokens so kernelName tiles == the old CSV-resolved tiles). pyflakes clean.
